% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/existing_youden.R
\name{existing_youden}
\alias{existing_youden}
\title{Find the threshold that maximizes the Youden Index by using only the existing values from the dataset as thresholds}
\usage{
existing_youden(roc.object, predictor)
}
\arguments{
\item{roc.object}{object that is defined by the roc() function output}

\item{predictor}{defined as dataset$predictor, predictor variable of the same dataset used in the roc.object}
}
\value{
Existing threshold value that maximizes the Youden Index is returned
}
\description{
Find the threshold that maximizes the Youden Index by using only the existing values from the dataset as thresholds
}
\examples{
existing_youden(ad.roc, AD$Tau)
}
